+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## unreleased
+
+- added missing clone field
+
 ## [v1.1.1] - 17.12.2019
 
 ### Fixed